The logged can bid on other peoples requests, He can also accept a bid that has been made on one of his posts.

+++ LETS/LETS/Controllers/HomeController.cs
@@ -131,7 +131,7 @@ namespace LETS.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult> PostUserBid(string username, int postId, float bid)
+        public async Task<ActionResult> PostUserBid(string username, int postId, int bid)
         {
             var user = username;
